pyproject.toml
[project]
name = "tasks-propig"
version = "0.1.0"
description = ""
authors = [
    {name = "itallo.minatti",email = "itallominatti@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "django (>=5.2.1,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "tasks_propig", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
